tracee-ebpf: the return of the pipeline
